11/21/2022
Added a bit more to constructors, expanded the inheritance chain so we wouldn't have to repeat as many methods, added Health and Strength Mechanics (Currently based on an illusive "Stage of the Game", will be expanded upon later). Added a LocationTemplate Class that will be expanded upon later, which builds locations, which will have child classes such as CityTemplate, and EnvironmentTemplate.